+++ %ProgramData%/Microsoft/Windows/UUS/State/_active.uusver
@@ -1,0 +1,1 @@
+923.508.111.0
