+++ App.js
@@ -1,10 +1,10 @@
 import { StatusBar } from "expo-status-bar";
 import { StyleSheet, Text, View } from "react-native";
 
+import Register from "./src/screens/RegisterScreen";
 import LoginScreen from "./src/screens/LoginScreen";
-import SignInScreen from "./src/screens/SignInScreen";
-import QrPay from "./src/screens/QrPay";
+import QrPay from "./src/screens/QrPayScreen";
 
 export default function App() {
-  return <SignInScreen />;
+  return <Register />;
 }
