+++ package.json
@@ -16,7 +16,8 @@
     "react-native": "0.71.7"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0"
+    "@babel/core": "^7.20.0",
+    "sass": "^1.62.1"
   },
   "private": true
 }
