+++ src/screens/QrPayScreen.js
